solved security vulnerabilities
I found a security vulnerability that let's me hijack flowers that are not mine due to the lack of permission checking in the giftFlower function - I can gift anyone any flower even though they are not mine. This problem is also found in the setForSale function - I can list and delist and flower for sale. I made use of an Owner modifier which checks for permission and it works well.